리팩토링 - App에서 List 컴포넌트 추출
- list 부분 컴포넌트로 추출
- App 테스트 수정
- List 테스트 추가
